Use f64 vector DVec3 and add random support

+++ src/camera.rs
@@ -30,7 +30,7 @@ impl Default for Camera {
 }
 
 impl Camera {
-  pub fn get_ray(&self, u: f32, v: f32) -> Ray {
+  pub fn get_ray(&self, u: f64, v: f64) -> Ray {
     Ray::new(
       &self.origin,
       &(self.lower_left_corner + u * self.horizontal + v * self.vertical - self.origin),
