+++ src/hittable.rs
@@ -5,7 +5,7 @@ use crate::ray::Ray;
 pub struct HitRecord {
   pub p: Point,
   pub normal: Vec3,
-  pub t: f32,
+  pub t: f64,
   pub front_face: bool,
 }
 
@@ -21,5 +21,5 @@ impl HitRecord {
 }
 
 pub trait Hittable: Send + Sync {
-  fn hit(&self, r: &Ray, t_min: f32, t_max: f32, rec: &mut HitRecord) -> bool;
+  fn hit(&self, r: &Ray, t_min: f64, t_max: f64, rec: &mut HitRecord) -> bool;
 }
